requirements.txt
GitPython==2.1.11
git-url-parse==1.2.2
semver==2.8.1
